refactor(package): move CLI into another package
Remove the CLI component of this package in favor of CLI consumers
installing a dedicated CLI package. By removing the CLI component, users
of this package who use it exclusively as a library do not need to
be burdened by the additional CLI dependencies.

BREAKING CHANGE:

Command line tool (CLI) has been removed from the
`@hutson/set-npm-auth-token-for-ci` package. Please use the
`@hutson/set-npm-auth-token-for-ci-cli` package if you would like to
continue to use the functionality of this package on the command line.